feat: Order 와 OrderItem 값 객체 테스트 및 제약 조건 추가
- 가격 제약 조건 : 가격은 0원부터 10억까지 값을 가진다.
- 주문 총 가격 제약 조건 : 주문 총 가격은 0원부터 10억 * 1000 까지의 값을 가진다.

+++ src/main/java/shopping/domain/order/Order.java
@@ -43,7 +43,7 @@ public class Order {
         orderPrice = orderPrice.plusPrice(orderItem.getTotalPrice());
     }
 
-    public int getOrderPrice() {
+    public long getOrderPrice() {
         return orderPrice.getValue();
     }
